src/main/java/com/xp/tdd/controller/CarController.java
package com.xp.tdd.controller;

import com.xp.tdd.domain.Car;
import com.xp.tdd.service.CarService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class CarController {

    private CarService carService;

    public CarController(CarService carService) {
        this.carService = carService;
    }

    @GetMapping("/cars/{name}")
    private Car getCar(@PathVariable String name){

        return carService.getCar(name);
    }
}
